# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AppointmentMode": {
        "enum": [
          "in_person",
          "teleconsult",
          "home_visit"
        ],
        "title": "AppointmentMode",
        "type": "string"
      },
      "AppointmentOut": {
        "properties": {
          "doctor_id": {
            "format": "uuid",
            "title": "Doctor Id",
            "type": "string"
          },
          "doctor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Name"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "follow_up_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "mode": {
            "$ref": "#/components/schemas/AppointmentMode"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          }
        },
        "required": [
          "id",
          "starts_at",
          "ends_at",
          "status",
          "mode",
          "reason",
          "location",
          "doctor_id",
          "doctor_name",
          "follow_up_id"
        ],
        "title": "AppointmentOut",
        "type": "object"
      },
      "AppointmentStatus": {
        "enum": [
          "requested",
          "scheduled",
          "confirmed",
          "checked_in",
          "completed",
          "cancelled",
          "no_show"
        ],
        "title": "AppointmentStatus",
        "type": "string"
      },
...
      "CaregiverDashboard": {
        "properties": {
          "people": {
            "items": {
              "$ref": "#/components/schemas/DependantSummary"
            },
            "title": "People",
            "type": "array"
          }
        },
        "required": [
          "people"
        ],
        "title": "CaregiverDashboard",
        "type": "object"
      },
...
      "CaregiverRelationshipType": {
        "enum": [
          "parent",
          "child",
          "spouse_partner",
          "sibling",
          "other_relative",
          "friend",
          "professional_carer",
          "legal_guardian",
          "other"
        ],
        "title": "CaregiverRelationshipType",
        "type": "string"
      },
...
      "DependantSummary": {
        "properties": {
          "appointments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AppointmentOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointments"
          },
          "follow_ups": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FollowUpOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Ups"
          },
          "is_dependant": {
            "title": "Is Dependant",
            "type": "boolean"
          },
          "is_guardian": {
            "title": "Is Guardian",
            "type": "boolean"
          },
          "missed_doses": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DoseOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missed Doses"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "recent_reports": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReportSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Reports"
          },
          "relationship_id": {
            "format": "uuid",
            "title": "Relationship Id",
            "type": "string"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/CaregiverRelationshipType"
          },
          "today_doses": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DoseOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Today Doses"
          }
        },
        "required": [
          "patient_id",
          "relationship_id",
          "name",
          "relationship_type",
          "is_guardian",
          "is_dependant",
          "permissions",
          "today_doses",
          "missed_doses",
          "appointments",
          "follow_ups",
          "recent_reports"
        ],
        "title": "DependantSummary",
        "type": "object"
      },
...
      "DoseOut": {
        "properties": {
          "as_needed": {
            "title": "As Needed",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "meal_relation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MealRelation"
              },
              {
                "type": "null"
              }
            ]
          },
          "medication_id": {
            "format": "uuid",
            "title": "Medication Id",
            "type": "string"
          },
          "medication_name": {
            "title": "Medication Name",
            "type": "string"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "skip_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Reason"
          },
          "snooze_count": {
            "title": "Snooze Count",
            "type": "integer"
          },
          "snoozed_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snoozed Until"
          },
          "source": {
            "$ref": "#/components/schemas/MedicationSource"
          },
          "status": {
            "$ref": "#/components/schemas/DoseStatus"
          },
          "strength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength"
          },
          "taken_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          }
        },
        "required": [
          "id",
          "medication_id",
          "medication_name",
          "strength",
          "instructions",
          "meal_relation",
          "source",
          "scheduled_at",
          "status",
          "snoozed_until",
          "snooze_count",
          "taken_at",
          "skip_reason",
          "as_needed"
        ],
        "title": "DoseOut",
        "type": "object"
      },
      "DoseStatus": {
        "enum": [
          "scheduled",
          "snoozed",
          "taken",
          "skipped",
          "missed",
          "cancelled"
        ],
        "title": "DoseStatus",
        "type": "string"
      },
...
      "FollowUpOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "doctor_id": {
            "format": "uuid",
            "title": "Doctor Id",
            "type": "string"
          },
          "doctor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Name"
          },
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "source_visit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Visit Id"
          },
          "status": {
            "$ref": "#/components/schemas/FollowUpStatus"
          }
        },
        "required": [
          "id",
          "due_date",
          "reason",
          "status",
          "doctor_id",
          "doctor_name",
          "source_visit_id",
          "completed_at"
        ],
        "title": "FollowUpOut",
        "type": "object"
      },
      "FollowUpStatus": {
        "enum": [
          "open",
          "booked",
          "completed",
          "cancelled"
        ],
        "title": "FollowUpStatus",
        "type": "string"
      },
...
      "MealRelation": {
        "enum": [
          "before_food",
          "after_food",
          "with_food",
          "empty_stomach",
          "bedtime",
          "any"
        ],
        "title": "MealRelation",
        "type": "string"
      },
...
      "MedicationSource": {
        "enum": [
          "prescription",
          "self_reported",
          "clinician_recorded",
          "integration"
        ],
        "title": "MedicationSource",
        "type": "string"
      },
...
      "ReportSummary": {
        "properties": {
          "collected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collected At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lab_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "lab_name",
          "source",
          "collected_at",
          "created_at",
          "result_count"
        ],
        "title": "ReportSummary",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /api/v1/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Health Io API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/caregiving/dashboard": {
      "get": {
        "operationId": "caregiver_dashboard_api_v1_me_caregiving_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverDashboard"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Caregiver Dashboard",
        "tags": [
          "caregivers"
        ]
      }
    },
...
  }
}